feat: add helper method for submission statuses
